refactor(uitests): split species-edit tests into SpeciesEditPanelUITests
CullingWorkflowUITests mixed toolbar/shortcut tests (01–30) with
species-edit-panel tests (41–54) in one shared-app class. test30's
Cmd+E export dialog occasionally didn't fully dismiss before test41
ran, and the resulting "ExifToggle not hittable" state leaked across
the rest of the 40s. The previous ci.yml worked around this by
running the two halves as separate xcodebuild invocations. After the
consolidation pass merged them back into one invocation the flake
re-surfaced.

Split the species-edit tests into their own class
(SpeciesEditPanelUITests) inheriting SuperPickyUITestCase. Each class
now gets its own app lifecycle via the base class's setUp/tearDown,
so test30 can't leak into test41 even when both classes run inside
one xcodebuild invocation. All panel-specific helpers (selectThumbnail,
arrowStripUntil, openPanelOnEagle, tapButton, focusSearchField, poll,
waitForSidecar, etc.) move with the tests — CullingWorkflowUITests
keeps only `dismissConfirmDialogIfPresent`, which is the only helper
its remaining tests (01–30) use.

Bonus: the split is now at the source level, so future sharding just
picks test classes — no need to enumerate test methods in ci.yml.
